requirements.txt
django==2.2.*
djangorestframework>=3.13.0,<3.14.0
psycopg2>=2.7.5,<2.8.0
flake8>=4.0.1,<4.1.0
pytest-cov